+++ skrub/_expressions/_utils.py
@@ -14,6 +14,7 @@ class Sentinels(enum.Enum):
     CROSS_VAL_DEFAULT_EXPR = "pipeline.expr"
     CROSS_VAL_DEFAULT_PIPELINE = "expr.skb.get_pipeline()"
     CROSS_VAL_DEFAULT_ENVIRONMENT = "expr.skb.get_data()"
+    OPTIONAL_VALUE = "value"
 
     def __repr__(self):
         return self.value
@@ -26,6 +27,7 @@ NULL = Sentinels.NULL
 CROSS_VAL_DEFAULT_EXPR = Sentinels.CROSS_VAL_DEFAULT_EXPR
 CROSS_VAL_DEFAULT_PIPELINE = Sentinels.CROSS_VAL_DEFAULT_PIPELINE
 CROSS_VAL_DEFAULT_ENVIRONMENT = Sentinels.CROSS_VAL_DEFAULT_ENVIRONMENT
+OPTIONAL_VALUE = Sentinels.OPTIONAL_VALUE
 
 
 def simple_repr(expr):